requirements.txt
telethon
supabase
python-telegram-bot
google-generativeai
python-dotenv
requests